# perf(inductor): improve `Adam` compile times drastically by moving list comprehensions into `_init_group`