@jon-chuang 2023-10-05
> This looks pretty good to me. Could you run the numbers a few times to have more statistically significant evidence?

Yes, I've run multiple times, sometimes inductor times change (it's quite variable) but dynamo times are quite invariant and improve by the same amount.