@jon-chuang 2023-10-05
> Ah I've discovered a problem--people call our functional APIs (adam, single_tensor_adam, multi_tensor_adam) and expect complex tensors to work. In this PR, that would no longer work :/

Right, that makes sense. In this case, we can do the conversion in loop using the alternate strategy, which is also faster.