@jon-chuang 2023-10-05
Superceded by: https://github.com/pytorch/pytorch/pull/110607